requirements.txt
grpcio==1.54.2 
grpcio-tools==1.54.2 
transformers==4.26.1
torch==1.13.1
protobuf
pdf2image
opencv-python
pytesseract
tqdm
huggingface_hub
openai
python-dotenv
sentencepiece
python-Levenshtein
fuzzywuzzy
tokenizers
fsspec
errant
setuptools